requirements.txt
slack-bolt==1.18.1
stripe==7.8.0
python-dotenv==1.0.0
flask==3.0.0
openai==1.3.7
pandas==2.1.4
numpy==1.25.2